20241119-python-web-development-with-django/django_agora/requirements.txt
asgiref==3.8.1
Django==5.1.3
django-multiselectfield==0.1.13
djangorestframework==3.15.2
pillow==11.0.0
psycopg2==2.9.10
sqlparse==0.5.2
